% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glaxo.R
\name{glaxo}
\alias{glaxo}
\title{Relaxed glasso}
\usage{
glaxo(Y, S = NULL, n = NULL, nlambda_relaxed = 4, ic = "bic", ...)
}
\arguments{
\item{Y}{A \code{data.frame} or matrix containing data. Specified if not supplying covariance matrix.}

\item{S}{A covariance matrix. Specified if not supplying raw data.}

\item{n}{Number of observations. Specified if not supplying raw data.}

\item{nlambda_relaxed}{Number of lambdas to be generated in Step 2 of algorithm}

\item{ic}{Information critera to be used for selecting model. See \code{\link[GGMncv]{ggmncv}} in the GGMncv package}

\item{...}{Other arguments to be passed on to \code{\link[GGMncv]{ggmncv}} in the GGMncv package}
}
\value{
An object of classes \code{glaxo}, \code{ggmncv}, and \code{default}
}
\description{
An implementation of relaxed glasso based on the 'simple algorithm'
for the relaxed lasso presented in section 2.2 of Meinshausen (2006)
}
\examples{
data("bfi", package = "psych")
Y <- subset(bfi, select = -c(gender, education, age))
Y <- na.omit(Y)

model <- glaxo(Y)
}
