% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plot_edges}
\alias{plot_edges}
\title{Plotting method for glaxo objects}
\usage{
plot_edges(
  object,
  layout = "fruchtermanreingold",
  node_labels = NULL,
  node_label_size = 5,
  node_size = 11,
  node_color = "white",
  node_border_size = 12,
  node_border_color = "black",
  edge_alpha = 1,
  ...
)
}
\arguments{
\item{object}{An object of type \code{glaxo}}

\item{layout}{A layout as specified in \code{\link[ggnetwork]{ggnetwork}}}

\item{node_labels}{Labels for nodes. Optional.}

\item{node_label_size}{Size of node labels. Default is 5.}

\item{node_size}{Size of nodes. Default is 11.}

\item{node_color}{Color of nodes. Default is "white".}

\item{node_border_size}{Size of node borders. Default is 12.}

\item{node_border_color}{Color of node borders. Default is "black".}

\item{edge_alpha}{Opaqueness of edges. Default is 1.}

\item{...}{Additional arguments to be passed to \code{\link[ggnetwork]{ggnetwork}}}
}
\description{
Plotting method for glaxo objects
}
\examples{
data("bfi", package = "psych")
Y <- subset(bfi, select = -c(gender, education, age))
Y <- na.omit(Y)

model <- glaxo(Y, n, progress = FALSE)

plot_edges(model)

}
