% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{predict.glaxo}
\alias{predict.glaxo}
\title{Predict method for glaxo object}
\usage{
\method{predict}{glaxo}(object, newdata, scale = TRUE)
}
\arguments{
\item{object}{An object of type \code{glaxo}}

\item{newdata}{A \code{data.frame} or matrix with with same number of columns as the original data.}

\item{scale}{Whether the newdata should be scaled. Default is true.}
}
\value{
A list containing the predictions and associated coefficients
\describe{
\item{predictions}{...}
\item{beta_matrix}{A \eqn{p x p}  matrix with beta coefficients on the off-diagonals and variances on the diagonals}
}
}
\description{
Generate predictions for new data based on glasso estimates
}
